ドキュメントのフォーマット(VS2005標準形式) 不要なusing句の削除 Loggerに渡すTypeの取得方法を統一 クラス変数名の命名規則の統一 クラス変数にreadonlyの付与
git-svn-id: https://www.seasar.org/svn/s2container.net/trunk@701 6374060a-8204-0410-be21-af3d0958e6ef

+++ s2container.net/source/Seasar.Examples/Seasar.Examples/Reference/ADO/UpdateClient.cs
@@ -23,24 +23,24 @@ using Seasar.Framework.Container.Factory;
 
 namespace Seasar.Examples.Reference.ADO
 {
-	public class UpdateClient
-	{
-		private const string PATH = "Seasar.Examples/Reference/ADO/Update.dicon";
+    public class UpdateClient
+    {
+        private const string PATH = "Seasar.Examples/Reference/ADO/Update.dicon";
 
-		public void Main()
-		{
-			IS2Container container = S2ContainerFactory.Create(PATH);
-			container.Init();
-			try
-			{
-				IUpdateHandler handler = (IUpdateHandler) container.GetComponent("UpdateHandler");
-				int result = (int) handler.Execute(new object[] { "SCOTT", 7788 });
-				Console.Out.WriteLine(result);
-			}
-			finally
-			{
-				container.Destroy();
-			}
-		}
-	}
+        public void Main()
+        {
+            IS2Container container = S2ContainerFactory.Create(PATH);
+            container.Init();
+            try
+            {
+                IUpdateHandler handler = (IUpdateHandler) container.GetComponent("UpdateHandler");
+                int result = handler.Execute(new object[] { "SCOTT", 7788 });
+                Console.Out.WriteLine(result);
+            }
+            finally
+            {
+                container.Destroy();
+            }
+        }
+    }
 }
